refactor(fhirfakes): rename Augmentor to Enricher per PR review
A reviewer flagged "Augmentor" as an odd name for the resource-graph
extension seam in Ignixa.FhirFakes.Workflow. Renamed to "Enricher"
(IResourceGraphEnricher, ResourceGraphEnrichmentContext,
AppointmentSchedulingEnricher, Enrich method) to avoid colliding with
existing Builder/Composer/Binder terminology and to use the standard
term for adding supplementary data to an existing object.
