+++ src/Core/Ignixa.FhirFakes/Workflow/IResourceGraphEnricher.cs
@@ -8,11 +8,11 @@ namespace Ignixa.FhirFakes.Workflow;
 /// <summary>
 /// Adds workflow resources (appointments, lists, document references, topology) to an existing
 /// resource graph. Implementations should be stateless with respect to execution: all per-run state
-/// lives on <see cref="ResourceGraph"/> or <see cref="ResourceGraphAugmentationContext"/> rather than
+/// lives on <see cref="ResourceGraph"/> or <see cref="ResourceGraphEnrichmentContext"/> rather than
 /// mutable instance fields, so a single configured instance is safe to reuse.
 /// </summary>
-public interface IResourceGraphAugmentor
+public interface IResourceGraphEnricher
 {
-    /// <summary>Mutates <paramref name="graph"/> in place, adding this augmentor's resources.</summary>
-    void Augment(ResourceGraph graph, ResourceGraphAugmentationContext context);
+    /// <summary>Mutates <paramref name="graph"/> in place, adding this enricher's resources.</summary>
+    void Enrich(ResourceGraph graph, ResourceGraphEnrichmentContext context);
 }
